Add new sendCommandCluster option (#223)
This adds a `sendCommandCluster` option as an alternative to `sendCommand`.

It can be set to a function that is passed an object with the key and isReadOnly value, in addition to the command. This provides everything that node-redis needs in cluster mode, and is expandable with new fields if necessary to support other libraries in cluster mode.

Additionally I tweaked the internals to always pass this information along - previously the key would be lost when loading the script, so it could inadvertenly get sent to the wrong redis server.

Fixes #222
Fixes #207
Fixes #208